+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed 
 
 - Exporting boreholes as CSV and JSON (without attachments) is now also available in anonymous mode. 
+- Display an error message to the user when attachment files could not be fetched from cloud storage during export.
 - Removed all asterisks from required form fields.
 - Updated the UI design of the tabs component.
 - Reincluded download link for codelist references in the import panel.